changed uint64 to uint32

+++ services/store/client/mock_client.go
@@ -149,7 +149,7 @@ func (m *MockStoreClient) GetObjectCustomMetadata(bucket string, objectKey strin
 	return nil
 }
 
-func (m *MockStoreClient) UpdateUploadStatus(uploadId string, status string, revision uint64, el *cerror.ErrorList) *proto.UpdateUploadStatusResponse {
+func (m *MockStoreClient) UpdateUploadStatus(uploadId string, status string, revision uint32, el *cerror.ErrorList) *proto.UpdateUploadStatusResponse {
 	args := m.Called(uploadId, status, revision, el)
 	if resp, ok := args.Get(0).(*proto.UpdateUploadStatusResponse); ok {
 		return resp
